backend/requirements.txt
fastapi[standard]
uvicorn
langchain-openai
langchain
langchain-community
python-dotenv
duckduckgo-search
psycopg2-binary
sqlalchemy